made some fixes on ubuntu

+++ src/dip/nodes/node_integer.cpp
@@ -89,37 +89,37 @@ namespace dip {
     // TODO: variable precision x should be implemented
     switch (value_dtype) {
     case val::DataType::Integer16_U: {
-      std::vector<unsigned short> arr;
+      std::vector<uint16_t> arr;
       arr.reserve(value_inputs.size());
       for (auto s: value_inputs) arr.push_back((unsigned short)std::stoul(s));
       return std::make_unique<val::ArrayValue<uint16_t>>(arr, shape);
     }
     case val::DataType::Integer16: {
-      std::vector<short> arr;
+      std::vector<int16_t> arr;
       arr.reserve(value_inputs.size());
       for (auto s: value_inputs) arr.push_back((short)std::stoi(s));
       return std::make_unique<val::ArrayValue<int16_t>>(arr, shape);
     }
     case val::DataType::Integer32_U: {
-      std::vector<unsigned int> arr;
+      std::vector<uint32_t> arr;
       arr.reserve(value_inputs.size());
       for (auto s: value_inputs) arr.push_back(std::stoul(s));
       return std::make_unique<val::ArrayValue<uint32_t>>(arr, shape);
     }
     case val::DataType::Integer32: {
-      std::vector<int> arr;
+      std::vector<int32_t> arr;
       arr.reserve(value_inputs.size());
       for (auto s: value_inputs) arr.push_back(std::stoi(s));
       return std::make_unique<val::ArrayValue<int32_t>>(arr, shape);
     }
     case val::DataType::Integer64_U: {
-      std::vector<unsigned long long> arr;
+      std::vector<uint64_t> arr;
       arr.reserve(value_inputs.size());
       for (auto s: value_inputs) arr.push_back(std::stoull(s));
       return std::make_unique<val::ArrayValue<uint64_t>>(arr, shape);
     }
     case val::DataType::Integer64: {
-      std::vector<long long> arr;
+      std::vector<int64_t> arr;
       arr.reserve(value_inputs.size());
       for (auto s: value_inputs) arr.push_back(std::stoll(s));
       return std::make_unique<val::ArrayValue<int64_t>>(arr, shape);
